+++ crates/engine/tree/src/tree/root.rs
@@ -4,7 +4,7 @@ use futures::Stream;
 use pin_project::pin_project;
 use reth_provider::providers::ConsistentDbView;
 use reth_trie::{updates::TrieUpdates, TrieInput};
-use reth_trie_parallel::parallel_root::ParallelStateRootError;
+use reth_trie_parallel::root::ParallelStateRootError;
 use revm_primitives::{EvmState, B256};
 use std::{
     future::Future,
